refactor(db): move floating functions to their own files
Fixes NODE-1255
