+++ lib/db.js
@@ -4,8 +4,6 @@ const EventEmitter = require('events').EventEmitter;
 const inherits = require('util').inherits;
 const getSingleProperty = require('./utils').getSingleProperty;
 const shallowClone = require('./utils').shallowClone;
-const parseIndexOptions = require('./utils').parseIndexOptions;
-const debugOptions = require('./utils').debugOptions;
 const CommandCursor = require('./command_cursor');
 const handleCallback = require('./utils').handleCallback;
 const filterOptions = require('./utils').filterOptions;
@@ -13,52 +11,33 @@ const toError = require('./utils').toError;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const f = require('util').format;
 const Admin = require('./admin');
-const Code = require('mongodb-core').BSON.Code;
 const MongoError = require('mongodb-core').MongoError;
 const ObjectID = require('mongodb-core').ObjectID;
 const Logger = require('mongodb-core').Logger;
 const Collection = require('./collection');
-const crypto = require('crypto');
 const mergeOptionsAndWriteConcern = require('./utils').mergeOptionsAndWriteConcern;
 const executeOperation = require('./utils').executeOperation;
 const applyWriteConcern = require('./utils').applyWriteConcern;
 
-var debugFields = [
-  'authSource',
-  'w',
-  'wtimeout',
-  'j',
-  'native_parser',
-  'forceServerObjectId',
-  'serializeFunctions',
-  'raw',
-  'promoteLongs',
-  'promoteValues',
-  'promoteBuffers',
-  'bufferMaxEntries',
-  'numberOfRetries',
-  'retryMiliSeconds',
-  'readPreference',
-  'pkFactory',
-  'parentDb',
-  'promiseLibrary',
-  'noListener'
-];
+// Operations
+const addUser = require('./operations/db_user_ops').addUser;
+const collections = require('./operations/db_collections');
+const createCollection = require('./operations/db_collection_ops').createCollection;
+const createIndex = require('./operations/db_index_ops').createIndex;
+const dropCollection = require('./operations/db_collection_ops').dropCollection;
+const dropDatabase = require('./operations/db_drop_database');
+const ensureIndex = require('./operations/db_index_ops').ensureIndex;
+const evaluate = require('./operations/db_evaluate');
+const executeCommand = require('./operations/db_execute_commands').executeCommand;
+const executeDbAdminCommand = require('./operations/db_execute_commands').executeDbAdminCommand;
+const indexInformation = require('./operations/db_index_ops').indexInformation;
+const removeUser = require('./operations/db_user_ops').removeUser;
 
-// Filter out any write concern options
-var illegalCommandFields = [
-  'w',
-  'wtimeout',
-  'j',
-  'fsync',
-  'autoIndexId',
-  'strict',
-  'serializeFunctions',
-  'pkFactory',
-  'raw',
-  'readPreference',
-  'session'
-];
+// Helpers
+const convertReadPreference = require('./operations/db_helpers').convertReadPreference;
+const profilingInfo = require('./operations/db_helpers').profilingInfo;
+const profilingLevel = require('./operations/db_helpers').profilingLevel;
+const setProfilingLevel = require('./operations/db_helpers').setProfilingLevel;
 
 /**
  * @fileOverview The **Db** class is a class that represents a MongoDB Database.
@@ -267,78 +246,6 @@ Object.defineProperty(Db.prototype, 'writeConcern', {
 });
 
 /**
- * Ensures provided read preference is properly converted into an object
- * @param {(ReadPreference|string|object)} readPreference the user provided read preference
- * @return {ReadPreference}
- */
-const convertReadPreference = function(readPreference) {
-  if (readPreference) {
-    if (typeof readPreference === 'string') {
-      return new ReadPreference(readPreference);
-    } else if (
-      readPreference &&
-      !(readPreference instanceof ReadPreference) &&
-      typeof readPreference === 'object'
-    ) {
-      const mode = readPreference.mode || readPreference.preference;
-      if (mode && typeof mode === 'string') {
-        return new ReadPreference(mode, readPreference.tags, {
-          maxStalenessSeconds: readPreference.maxStalenessSeconds
-        });
-      }
-    } else if (!(readPreference instanceof ReadPreference)) {
-      throw new TypeError('Invalid read preference: ' + readPreference);
-    }
-  }
-
-  return readPreference;
-};
-
-/**
- * The callback format for results
- * @callback Db~resultCallback
- * @param {MongoError} error An error instance representing the error during the execution.
- * @param {object} result The result object if the command was executed successfully.
- */
-var executeCommand = function(self, command, options, callback) {
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed())
-    return callback(new MongoError('topology was destroyed'));
-  // Get the db name we are executing against
-  var dbName = options.dbName || options.authdb || self.s.databaseName;
-
-  // If we have a readPreference set
-  if (options.readPreference == null && self.s.readPreference) {
-    options.readPreference = self.s.readPreference;
-  }
-
-  // Convert the readPreference if its not a write
-  if (options.readPreference) {
-    options.readPreference = convertReadPreference(options.readPreference);
-  } else {
-    options.readPreference = ReadPreference.primary;
-  }
-
-  // Debug information
-  if (self.s.logger.isDebug())
-    self.s.logger.debug(
-      f(
-        'executing command %s against %s with options [%s]',
-        JSON.stringify(command),
-        f('%s.$cmd', dbName),
-        JSON.stringify(debugOptions(debugFields, options))
-      )
-    );
-
-  // Execute command
-  self.s.topology.command(f('%s.$cmd', dbName), command, options, function(err, result) {
-    if (err) return handleCallback(callback, err);
-    if (options.full) return handleCallback(callback, null, result);
-    handleCallback(callback, null, result.result);
-  });
-};
-
-/**
  * Execute a command
  * @method
  * @param {object} command The command hash
@@ -477,90 +384,6 @@ Db.prototype.collection = function(name, options, callback) {
       return handleCallback(callback, err, null);
     }
   });
-};
-
-var createCollection = function(self, name, options, callback) {
-  // Get the write concern options
-  const finalOptions = applyWriteConcern(shallowClone(options), { db: self }, options);
-
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed()) {
-    return callback(new MongoError('topology was destroyed'));
-  }
-
-  const listCollectionOptions = Object.assign({}, finalOptions, { nameOnly: true });
-
-  // Check if we have the name
-  self
-    .listCollections({ name: name }, listCollectionOptions)
-    .setReadPreference(ReadPreference.PRIMARY)
-    .toArray(function(err, collections) {
-      if (err != null) return handleCallback(callback, err, null);
-      if (collections.length > 0 && finalOptions.strict) {
-        return handleCallback(
-          callback,
-          MongoError.create({
-            message: f('Collection %s already exists. Currently in strict mode.', name),
-            driver: true
-          }),
-          null
-        );
-      } else if (collections.length > 0) {
-        try {
-          return handleCallback(
-            callback,
-            null,
-            new Collection(
-              self,
-              self.s.topology,
-              self.s.databaseName,
-              name,
-              self.s.pkFactory,
-              options
-            )
-          );
-        } catch (err) {
-          return handleCallback(callback, err);
-        }
-      }
-
-      // Create collection command
-      var cmd = { create: name };
-
-      // Decorate command with writeConcern if supported
-      applyWriteConcern(cmd, { db: self }, options);
-
-      // Add all optional parameters
-      for (var n in options) {
-        if (
-          options[n] != null &&
-          typeof options[n] !== 'function' &&
-          illegalCommandFields.indexOf(n) === -1
-        ) {
-          cmd[n] = options[n];
-        }
-      }
-
-      // Force a primary read Preference
-      finalOptions.readPreference = ReadPreference.PRIMARY;
-
-      // Execute command
-      self.command(cmd, finalOptions, function(err) {
-        if (err) return handleCallback(callback, err);
-        handleCallback(
-          callback,
-          null,
-          new Collection(
-            self,
-            self.s.topology,
-            self.s.databaseName,
-            name,
-            self.s.pkFactory,
-            options
-          )
-        );
-      });
-    });
 };
 
 /**
@@ -735,47 +558,6 @@ Db.prototype.listCollections = function(filter, options) {
   return cursor;
 };
 
-var evaluate = function(self, code, parameters, options, callback) {
-  var finalCode = code;
-  var finalParameters = [];
-
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed())
-    return callback(new MongoError('topology was destroyed'));
-
-  // If not a code object translate to one
-  if (!(finalCode && finalCode._bsontype === 'Code')) finalCode = new Code(finalCode);
-  // Ensure the parameters are correct
-  if (parameters != null && !Array.isArray(parameters) && typeof parameters !== 'function') {
-    finalParameters = [parameters];
-  } else if (parameters != null && Array.isArray(parameters) && typeof parameters !== 'function') {
-    finalParameters = parameters;
-  }
-
-  // Create execution selector
-  var cmd = { $eval: finalCode, args: finalParameters };
-  // Check if the nolock parameter is passed in
-  if (options['nolock']) {
-    cmd['nolock'] = options['nolock'];
-  }
-
-  // Set primary read preference
-  options.readPreference = new ReadPreference(ReadPreference.PRIMARY);
-
-  // Execute the command
-  self.command(cmd, options, function(err, result) {
-    if (err) return handleCallback(callback, err, null);
-    if (result && result.ok === 1) return handleCallback(callback, null, result.retval);
-    if (result)
-      return handleCallback(
-        callback,
-        MongoError.create({ message: f('eval failed: %s', result.errmsg), driver: true }),
-        null
-      );
-    handleCallback(callback, err, result);
-  });
-};
-
 /**
  * Evaluate JavaScript on the server
  *
@@ -851,19 +633,6 @@ Db.prototype.dropCollection = function(name, options, callback) {
   return executeOperation(this.s.topology, dropCollection, [this, cmd, opts, callback]);
 };
 
-const dropCollection = (self, cmd, options, callback) => {
-  return self.command(cmd, options, function(err, result) {
-    // Did the user destroy the topology
-    if (self.serverConfig && self.serverConfig.isDestroyed()) {
-      return callback(new MongoError('topology was destroyed'));
-    }
-
-    if (err) return handleCallback(callback, err);
-    if (result.ok) return handleCallback(callback, null, true);
-    handleCallback(callback, null, false);
-  });
-};
-
 /**
  * Drop a database, removing it permanently from the server.
  *
@@ -894,19 +663,6 @@ Db.prototype.dropDatabase = function(options, callback) {
   return executeOperation(this.s.topology, dropDatabase, [this, cmd, finalOptions, callback]);
 };
 
-const dropDatabase = (self, cmd, options, callback) => {
-  self.command(cmd, options, function(err, result) {
-    // Did the user destroy the topology
-    if (self.serverConfig && self.serverConfig.isDestroyed()) {
-      return callback(new MongoError('topology was destroyed'));
-    }
-
-    if (callback == null) return;
-    if (err) return handleCallback(callback, err, null);
-    handleCallback(callback, null, result.ok ? true : false);
-  });
-};
-
 /**
  * Fetch all collections for the current db.
  *
@@ -921,34 +677,6 @@ Db.prototype.collections = function(options, callback) {
   options = options || {};
 
   return executeOperation(this.s.topology, collections, [this, options, callback]);
-};
-
-var collections = function(self, options, callback) {
-  options = Object.assign({}, options, { nameOnly: true });
-  // Let's get the collection names
-  self.listCollections({}, options).toArray(function(err, documents) {
-    if (err != null) return handleCallback(callback, err, null);
-    // Filter collections removing any illegal ones
-    documents = documents.filter(function(doc) {
-      return doc.name.indexOf('$') === -1;
-    });
-
-    // Return the collection objects
-    handleCallback(
-      callback,
-      null,
-      documents.map(function(d) {
-        return new Collection(
-          self,
-          self.s.topology,
-          self.s.databaseName,
-          d.name,
-          self.s.pkFactory,
-          self.s.options
-        );
-      })
-    );
-  });
 };
 
 /**
@@ -976,18 +704,6 @@ Db.prototype.executeDbAdminCommand = function(selector, options, callback) {
     options,
     callback
   ]);
-};
-
-const executeDbAdminCommand = (self, selector, options, callback) => {
-  self.s.topology.command('admin.$cmd', selector, options, function(err, result) {
-    // Did the user destroy the topology
-    if (self.serverConfig && self.serverConfig.isDestroyed()) {
-      return callback(new MongoError('topology was destroyed'));
-    }
-
-    if (err) return handleCallback(callback, err);
-    handleCallback(callback, null, result.result);
-  });
 };
 
 /**
@@ -1026,65 +742,6 @@ Db.prototype.createIndex = function(name, fieldOrSpec, options, callback) {
   ]);
 };
 
-var createIndex = function(self, name, fieldOrSpec, options, callback) {
-  // Get the write concern options
-  var finalOptions = Object.assign({}, { readPreference: ReadPreference.PRIMARY }, options);
-  finalOptions = applyWriteConcern(finalOptions, { db: self }, options);
-
-  // Ensure we have a callback
-  if (finalOptions.writeConcern && typeof callback !== 'function') {
-    throw MongoError.create({
-      message: 'Cannot use a writeConcern without a provided callback',
-      driver: true
-    });
-  }
-
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed())
-    return callback(new MongoError('topology was destroyed'));
-
-  // Attempt to run using createIndexes command
-  createIndexUsingCreateIndexes(self, name, fieldOrSpec, options, function(err, result) {
-    if (err == null) return handleCallback(callback, err, result);
-
-    // 67 = 'CannotCreateIndex' (malformed index options)
-    // 85 = 'IndexOptionsConflict' (index already exists with different options)
-    // 86 = 'IndexKeySpecsConflict' (index already exists with the same name)
-    // 11000 = 'DuplicateKey' (couldn't build unique index because of dupes)
-    // 11600 = 'InterruptedAtShutdown' (interrupted at shutdown)
-    // These errors mean that the server recognized `createIndex` as a command
-    // and so we don't need to fallback to an insert.
-    if (
-      err.code === 67 ||
-      err.code === 11000 ||
-      err.code === 85 ||
-      err.code === 86 ||
-      err.code === 11600
-    ) {
-      return handleCallback(callback, err, result);
-    }
-
-    // Create command
-    var doc = createCreateIndexCommand(self, name, fieldOrSpec, options);
-    // Set no key checking
-    finalOptions.checkKeys = false;
-    // Insert document
-    self.s.topology.insert(
-      f('%s.%s', self.s.databaseName, Db.SYSTEM_INDEX_COLLECTION),
-      doc,
-      finalOptions,
-      function(err, result) {
-        if (callback == null) return;
-        if (err) return handleCallback(callback, err);
-        if (result == null) return handleCallback(callback, null, null);
-        if (result.result.writeErrors)
-          return handleCallback(callback, MongoError.create(result.result.writeErrors[0]), null);
-        handleCallback(callback, null, doc.name);
-      }
-    );
-  });
-};
-
 /**
  * Ensures that an index exists, if it does not it creates it
  * @method
@@ -1121,184 +778,9 @@ Db.prototype.ensureIndex = function(name, fieldOrSpec, options, callback) {
   ]);
 };
 
-var ensureIndex = function(self, name, fieldOrSpec, options, callback) {
-  // Get the write concern options
-  var finalOptions = applyWriteConcern({}, { db: self }, options);
-  // Create command
-  var selector = createCreateIndexCommand(self, name, fieldOrSpec, options);
-  var index_name = selector.name;
-
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed())
-    return callback(new MongoError('topology was destroyed'));
-
-  // Merge primary readPreference
-  finalOptions.readPreference = ReadPreference.PRIMARY;
-
-  // Check if the index allready exists
-  self.indexInformation(name, finalOptions, function(err, indexInformation) {
-    if (err != null && err.code !== 26) return handleCallback(callback, err, null);
-    // If the index does not exist, create it
-    if (indexInformation == null || !indexInformation[index_name]) {
-      self.createIndex(name, fieldOrSpec, options, callback);
-    } else {
-      if (typeof callback === 'function') return handleCallback(callback, null, index_name);
-    }
-  });
-};
-
 Db.prototype.addChild = function(db) {
   if (this.s.parentDb) return this.s.parentDb.addChild(db);
   this.s.children.push(db);
-};
-
-var _executeAuthCreateUserCommand = function(self, username, password, options, callback) {
-  // Special case where there is no password ($external users)
-  if (typeof username === 'string' && password != null && typeof password === 'object') {
-    options = password;
-    password = null;
-  }
-
-  // Unpack all options
-  if (typeof options === 'function') {
-    callback = options;
-    options = {};
-  }
-
-  // Error out if we digestPassword set
-  if (options.digestPassword != null) {
-    throw toError(
-      "The digestPassword option is not supported via add_user. Please use db.command('createUser', ...) instead for this option."
-    );
-  }
-
-  // Get additional values
-  var customData = options.customData != null ? options.customData : {};
-  var roles = Array.isArray(options.roles) ? options.roles : [];
-  var maxTimeMS = typeof options.maxTimeMS === 'number' ? options.maxTimeMS : null;
-
-  // If not roles defined print deprecated message
-  if (roles.length === 0) {
-    console.log('Creating a user without roles is deprecated in MongoDB >= 2.6');
-  }
-
-  // Get the error options
-  var commandOptions = { writeCommand: true };
-  if (options['dbName']) commandOptions.dbName = options['dbName'];
-
-  // Add maxTimeMS to options if set
-  if (maxTimeMS != null) commandOptions.maxTimeMS = maxTimeMS;
-
-  // Check the db name and add roles if needed
-  if (
-    (self.databaseName.toLowerCase() === 'admin' || options.dbName === 'admin') &&
-    !Array.isArray(options.roles)
-  ) {
-    roles = ['root'];
-  } else if (!Array.isArray(options.roles)) {
-    roles = ['dbOwner'];
-  }
-
-  const digestPassword = self.s.topology.lastIsMaster().maxWireVersion >= 7;
-
-  // Build the command to execute
-  var command = {
-    createUser: username,
-    customData: customData,
-    roles: roles,
-    digestPassword
-  };
-
-  // Apply write concern to command
-  command = applyWriteConcern(command, { db: self }, options);
-
-  let userPassword = password;
-
-  if (!digestPassword) {
-    // Use node md5 generator
-    let md5 = crypto.createHash('md5');
-    // Generate keys used for authentication
-    md5.update(username + ':mongo:' + password);
-    userPassword = md5.digest('hex');
-  }
-
-  // No password
-  if (typeof password === 'string') {
-    command.pwd = userPassword;
-  }
-
-  // Force write using primary
-  commandOptions.readPreference = ReadPreference.primary;
-
-  // Execute the command
-  self.command(command, commandOptions, function(err, result) {
-    if (err && err.ok === 0 && err.code === undefined)
-      return handleCallback(callback, { code: -5000 }, null);
-    if (err) return handleCallback(callback, err, null);
-    handleCallback(
-      callback,
-      !result.ok ? toError(result) : null,
-      result.ok ? [{ user: username, pwd: '' }] : null
-    );
-  });
-};
-
-var addUser = function(self, username, password, options, callback) {
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed())
-    return callback(new MongoError('topology was destroyed'));
-  // Attempt to execute auth command
-  _executeAuthCreateUserCommand(self, username, password, options, function(err, r) {
-    // We need to perform the backward compatible insert operation
-    if (err && err.code === -5000) {
-      var finalOptions = applyWriteConcern(shallowClone(options), { db: self }, options);
-
-      // Use node md5 generator
-      var md5 = crypto.createHash('md5');
-      // Generate keys used for authentication
-      md5.update(username + ':mongo:' + password);
-      var userPassword = md5.digest('hex');
-
-      // If we have another db set
-      var db = options.dbName ? new Db(options.dbName, self.s.topology, self.s.options) : self;
-
-      // Fetch a user collection
-      var collection = db.collection(Db.SYSTEM_USER_COLLECTION);
-
-      // Check if we are inserting the first user
-      collection.count({}, finalOptions, function(err, count) {
-        // We got an error (f.ex not authorized)
-        if (err != null) return handleCallback(callback, err, null);
-        // Check if the user exists and update i
-        collection
-          .find({ user: username }, { dbName: options['dbName'] }, finalOptions)
-          .toArray(function(err) {
-            // We got an error (f.ex not authorized)
-            if (err != null) return handleCallback(callback, err, null);
-            // Add command keys
-            finalOptions.upsert = true;
-
-            // We have a user, let's update the password or upsert if not
-            collection.update(
-              { user: username },
-              { $set: { user: username, pwd: userPassword } },
-              finalOptions,
-              function(err) {
-                if (count === 0 && err)
-                  return handleCallback(callback, null, [{ user: username, pwd: userPassword }]);
-                if (err) return handleCallback(callback, err, null);
-                handleCallback(callback, null, [{ user: username, pwd: userPassword }]);
-              }
-            );
-          });
-      });
-
-      return;
-    }
-
-    if (err) return handleCallback(callback, err);
-    handleCallback(callback, err, r);
-  });
 };
 
 /**
@@ -1323,69 +805,6 @@ Db.prototype.addUser = function(username, password, options, callback) {
   options = args.length ? args.shift() || {} : {};
 
   return executeOperation(this.s.topology, addUser, [this, username, password, options, callback]);
-};
-
-var _executeAuthRemoveUserCommand = function(self, username, options, callback) {
-  if (typeof options === 'function') (callback = options), (options = {});
-  options = options || {};
-
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed())
-    return callback(new MongoError('topology was destroyed'));
-  // Get the error options
-  var commandOptions = { writeCommand: true };
-  if (options['dbName']) commandOptions.dbName = options['dbName'];
-
-  // Get additional values
-  var maxTimeMS = typeof options.maxTimeMS === 'number' ? options.maxTimeMS : null;
-
-  // Add maxTimeMS to options if set
-  if (maxTimeMS != null) commandOptions.maxTimeMS = maxTimeMS;
-
-  // Build the command to execute
-  var command = {
-    dropUser: username
-  };
-
-  // Apply write concern to command
-  command = applyWriteConcern(command, { db: self }, options);
-
-  // Force write using primary
-  commandOptions.readPreference = ReadPreference.primary;
-
-  // Execute the command
-  self.command(command, commandOptions, function(err, result) {
-    if (err && !err.ok && err.code === undefined) return handleCallback(callback, { code: -5000 });
-    if (err) return handleCallback(callback, err, null);
-    handleCallback(callback, null, result.ok ? true : false);
-  });
-};
-
-var removeUser = function(self, username, options, callback) {
-  // Attempt to execute command
-  _executeAuthRemoveUserCommand(self, username, options, function(err, result) {
-    if (err && err.code === -5000) {
-      var finalOptions = applyWriteConcern(shallowClone(options), { db: self }, options);
-      // If we have another db set
-      var db = options.dbName ? new Db(options.dbName, self.s.topology, self.s.options) : self;
-
-      // Fetch a user collection
-      var collection = db.collection(Db.SYSTEM_USER_COLLECTION);
-
-      // Locate the user
-      collection.findOne({ user: username }, finalOptions, function(err, user) {
-        if (user == null) return handleCallback(callback, err, false);
-        collection.remove({ user: username }, finalOptions, function(err) {
-          handleCallback(callback, err, true);
-        });
-      });
-
-      return;
-    }
-
-    if (err) return handleCallback(callback, err);
-    handleCallback(callback, err, result);
-  });
 };
 
 /**
@@ -1424,31 +843,6 @@ Db.prototype.setProfilingLevel = function(level, options, callback) {
   return executeOperation(this.s.topology, setProfilingLevel, [this, level, options, callback]);
 };
 
-var setProfilingLevel = function(self, level, options, callback) {
-  var command = {};
-  var profile = 0;
-
-  if (level === 'off') {
-    profile = 0;
-  } else if (level === 'slow_only') {
-    profile = 1;
-  } else if (level === 'all') {
-    profile = 2;
-  } else {
-    return callback(new Error('Error: illegal profiling level value ' + level));
-  }
-
-  // Set up the profile number
-  command['profile'] = profile;
-
-  self.command(command, options, function(err, doc) {
-    if (err == null && doc.ok === 1) return callback(null, level);
-    return err != null
-      ? callback(err, null)
-      : callback(new Error('Error with profile command'), null);
-  });
-};
-
 /**
  * Retrive the current profiling information for MongoDB
  *
@@ -1465,17 +859,6 @@ Db.prototype.profilingInfo = function(options, callback) {
   return executeOperation(this.s.topology, profilingInfo, [this, options, callback]);
 };
 
-var profilingInfo = function(self, options, callback) {
-  try {
-    self
-      .collection('system.profile')
-      .find({}, null, options)
-      .toArray(callback);
-  } catch (err) {
-    return callback(err, null);
-  }
-};
-
 /**
  * Retrieve the current profiling Level for MongoDB
  *
@@ -1489,20 +872,6 @@ Db.prototype.profilingLevel = function(options, callback) {
   options = options || {};
 
   return executeOperation(this.s.topology, profilingLevel, [this, options, callback]);
-};
-
-var profilingLevel = function(self, options, callback) {
-  self.command({ profile: -1 }, options, function(err, doc) {
-    if (err == null && doc.ok === 1) {
-      var was = doc.was;
-      if (was === 0) return callback(null, 'off');
-      if (was === 1) return callback(null, 'slow_only');
-      if (was === 2) return callback(null, 'all');
-      return callback(new Error('Error: illegal profiling level value ' + was), null);
-    } else {
-      err != null ? callback(err, null) : callback(new Error('Error with profile command'), null);
-    }
-  });
 };
 
 /**
@@ -1521,129 +890,6 @@ Db.prototype.indexInformation = function(name, options, callback) {
   options = options || {};
 
   return executeOperation(this.s.topology, indexInformation, [this, name, options, callback]);
-};
-
-var indexInformation = function(self, name, options, callback) {
-  // If we specified full information
-  var full = options['full'] == null ? false : options['full'];
-
-  // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed())
-    return callback(new MongoError('topology was destroyed'));
-  // Process all the results from the index command and collection
-  var processResults = function(indexes) {
-    // Contains all the information
-    var info = {};
-    // Process all the indexes
-    for (var i = 0; i < indexes.length; i++) {
-      var index = indexes[i];
-      // Let's unpack the object
-      info[index.name] = [];
-      for (var name in index.key) {
-        info[index.name].push([name, index.key[name]]);
-      }
-    }
-
-    return info;
-  };
-
-  // Get the list of indexes of the specified collection
-  self
-    .collection(name)
-    .listIndexes(options)
-    .toArray(function(err, indexes) {
-      if (err) return callback(toError(err));
-      if (!Array.isArray(indexes)) return handleCallback(callback, null, []);
-      if (full) return handleCallback(callback, null, indexes);
-      handleCallback(callback, null, processResults(indexes));
-    });
-};
-
-var createCreateIndexCommand = function(db, name, fieldOrSpec, options) {
-  var indexParameters = parseIndexOptions(fieldOrSpec);
-  var fieldHash = indexParameters.fieldHash;
-
-  // Generate the index name
-  var indexName = typeof options.name === 'string' ? options.name : indexParameters.name;
-  var selector = {
-    ns: db.databaseName + '.' + name,
-    key: fieldHash,
-    name: indexName
-  };
-
-  // Ensure we have a correct finalUnique
-  var finalUnique = options == null || 'object' === typeof options ? false : options;
-  // Set up options
-  options = options == null || typeof options === 'boolean' ? {} : options;
-
-  // Add all the options
-  var keysToOmit = Object.keys(selector);
-  for (var optionName in options) {
-    if (keysToOmit.indexOf(optionName) === -1) {
-      selector[optionName] = options[optionName];
-    }
-  }
-
-  if (selector['unique'] == null) selector['unique'] = finalUnique;
-
-  // Remove any write concern operations
-  var removeKeys = ['w', 'wtimeout', 'j', 'fsync', 'readPreference'];
-  for (var i = 0; i < removeKeys.length; i++) {
-    delete selector[removeKeys[i]];
-  }
-
-  // Return the command creation selector
-  return selector;
-};
-
-var createIndexUsingCreateIndexes = function(self, name, fieldOrSpec, options, callback) {
-  // Build the index
-  var indexParameters = parseIndexOptions(fieldOrSpec);
-  // Generate the index name
-  var indexName = typeof options.name === 'string' ? options.name : indexParameters.name;
-  // Set up the index
-  var indexes = [{ name: indexName, key: indexParameters.fieldHash }];
-  // merge all the options
-  var keysToOmit = Object.keys(indexes[0]).concat([
-    'w',
-    'wtimeout',
-    'j',
-    'fsync',
-    'readPreference',
-    'session'
-  ]);
-
-  for (var optionName in options) {
-    if (keysToOmit.indexOf(optionName) === -1) {
-      indexes[0][optionName] = options[optionName];
-    }
-  }
-
-  // Get capabilities
-  var capabilities = self.s.topology.capabilities();
-
-  // Did the user pass in a collation, check if our write server supports it
-  if (indexes[0].collation && capabilities && !capabilities.commandsTakeCollation) {
-    // Create a new error
-    var error = new MongoError(f('server/primary/mongos does not support collation'));
-    error.code = 67;
-    // Return the error
-    return callback(error);
-  }
-
-  // Create command, apply write concern to command
-  var cmd = applyWriteConcern({ createIndexes: name, indexes: indexes }, { db: self }, options);
-
-  // ReadPreference primary
-  options.readPreference = ReadPreference.PRIMARY;
-
-  // Build the command
-  self.command(cmd, options, function(err, result) {
-    if (err) return handleCallback(callback, err, null);
-    if (result.ok === 0) return handleCallback(callback, toError(result), null);
-    // Return the indexName for backward compatibility
-    handleCallback(callback, null, indexName);
-  });
 };
 
 // Validate the database name
